clean up scripts folder
move legacy
