+++ legacy/commands.sh
