+++ legacy/selector.js
